🚚 move: NavBar device 폴더 분리
