+++ app/src/components/elements/NavBar/Desktop/index.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { NavBar } from './NavBar';
+import { NavBar } from '../shared/NavBar';
 
 export const DesktopNavBar = () => {
   return (
