+++ app/src/components/elements/NavBar/shared/NavItem.tsx
@@ -1,7 +1,7 @@
 import { HStack, Text } from '@components/common';
 import styled from '@emotion/styled';
 import { useLocation, useNavigate } from 'react-router-dom';
-import type { NavRoute } from './hooks';
+import type { NavRoute } from '../hooks';
 
 export type NavItemProps = {
   route: NavRoute;
