+++ app/src/components/elements/NavBar/shared/NavMenu.tsx
@@ -1,6 +1,6 @@
 import { Spacer, VStack } from '@components/common';
 import { NavItem } from './NavItem';
-import { useNavRoutes } from './hooks';
+import { useNavRoutes } from '../hooks';
 
 // TODO: SidebarMenu와 TabBar의 로직 유사하므로 통합 필요
 export const NavMenu = () => {
