+++ app/src/components/elements/NavBar/shared/NavProfile.tsx
