+++ app/src/components/layouts/MainLayout/DesktopMainLayout.tsx
@@ -1,6 +1,6 @@
 import { HStack, VStack } from '@components/common';
 import { IntraLink } from '@components/elements/IntraLink';
-import { DesktopNavBar } from '@components/elements/NavBar/DesktopNavBar';
+import { DesktopNavBar } from '@components/elements/NavBar/Desktop';
 import styled from '@emotion/styled';
 import { Outlet } from 'react-router-dom';
