+++ app/src/components/layouts/MainLayout/MobileMainLayout.tsx
@@ -1,7 +1,7 @@
 import { VStack } from '@components/common';
 import { MobileHeader } from '@components/elements/Header/Mobile';
 import { IntraLink } from '@components/elements/IntraLink';
-import { TabBar } from '@components/elements/NavBar/TabBar';
+import { TabBar } from '@components/elements/NavBar/Mobile(TabBar)';
 import { useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
 import { Helmet } from 'react-helmet-async';
