+++ app/src/components/layouts/MainLayout/TabletMainLayout.tsx
@@ -1,6 +1,6 @@
 import { HStack, VStack } from '@components/common';
 import { IntraLink } from '@components/elements/IntraLink';
-import { TabletNavBar } from '@components/elements/NavBar/TabletNavBar';
+import { TabletNavBar } from '@components/elements/NavBar/Tablet';
 import styled from '@emotion/styled';
 import { Outlet } from 'react-router-dom';
